refactor: simplify molecule page routes
